@@ -3,7 +3,7 @@ package config
 // NewMockConfig returns a config to be used in tests
 func NewMockConfig() *Config {
 	cfg := new(Config)
-	cfg.ServicePort = 9000
+	cfg.BindPort = 9000
 	cfg.Certificate = "/path/cert.pem"
 	cfg.CertificateKey = "/path/certkey.pem"
 	cfg.CertificateAuthoritiesDir = "/path/to/cas"
